django-quiz-poll/requirements/base.txt
Django==1.11
djangorestframework==3.7
mimesis
mysqlclient
flake8
